Tools for visualizing job-running
